+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), following [Se
 
 ## [Unreleased]
 
+## [0.4.0]
+
 - change: `app_factory_func` now checks if the passed App deploys successfully. This is configured using
   the `timeout_sec` argument, which by default is equal 60. Use value 0 to disable checking if App deploys OK.
 - added: